separate profile uploads spec into its own file
